FewPalettes refactoring
