+++ data/src/commonMain/kotlin/dev/t1r/themebuilder/data/colors/material/MaterialColorsRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package dev.t1r.themebuilder.data.colors.material
 
 import dev.t1r.themebuilder.entity.colors.ColorGroup
+import dev.t1r.themebuilder.repository.colors.material.MaterialColorsRepository
 
 class MaterialColorsRepositoryImpl(
     private val dataSource: MaterialColorsDataSource,
