+++ entity/build.gradle.kts
@@ -6,6 +6,7 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
+                implementation(Deps.Kotlin.coroutines)
             }
         }
     }
