+++ feature-baseline-colors/src/commonMain/kotlin/dev/t1r/themebuilder/feature/baselinecolor/integration/BaselineColorsComponentImpl.kt
@@ -3,10 +3,10 @@ package dev.t1r.themebuilder.feature.baselinecolor.integration
 import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.mvikotlin.core.store.StoreFactory
 import com.arkivanov.mvikotlin.extensions.coroutines.states
-import dev.t1r.themebuilder.data.colors.theme.ThemeColorsRepository
 import dev.t1r.themebuilder.feature.baselinecolor.BaselineColorsComponent
 import dev.t1r.themebuilder.feature.baselinecolor.BaselineColorsComponent.Model
 import dev.t1r.themebuilder.feature.baselinecolor.store.BaselineColorsStoreProvider
+import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
