+++ feature-baseline-colors/src/commonMain/kotlin/dev/t1r/themebuilder/feature/baselinecolor/store/BaselineColorsStoreProvider.kt
@@ -1,13 +1,13 @@
 package dev.t1r.themebuilder.feature.baselinecolor.store
 
 import com.arkivanov.mvikotlin.core.store.Reducer
-import com.arkivanov.mvikotlin.core.store.StoreFactory
 import com.arkivanov.mvikotlin.core.store.Store
+import com.arkivanov.mvikotlin.core.store.StoreFactory
 import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineBootstrapper
 import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineExecutor
-import dev.t1r.themebuilder.data.colors.theme.ThemeColorsRepository
 import dev.t1r.themebuilder.entity.colors.ThemeColors
-import  dev.t1r.themebuilder.feature.baselinecolor.store.BaselineColorsStore.*
+import dev.t1r.themebuilder.feature.baselinecolor.store.BaselineColorsStore.*
+import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
