+++ feature-material-colors-pallet/src/commonMain/kotlin/dev/t1r/themebuilder/feature/materialcolorspallet/integration/MaterialColorsPalletComponentImpl.kt
@@ -3,14 +3,14 @@ package dev.t1r.themebuilder.feature.materialcolorspallet.integration
 import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.mvikotlin.core.store.StoreFactory
 import com.arkivanov.mvikotlin.extensions.coroutines.states
-import dev.t1r.themebuilder.data.colors.material.MaterialColorsRepository
-import dev.t1r.themebuilder.data.colors.theme.ThemeColorsRepository
 import dev.t1r.themebuilder.entity.colors.ColorModel
 import dev.t1r.themebuilder.entity.colors.ThemeColorsEnum
 import dev.t1r.themebuilder.feature.materialcolorspallet.MaterialColorsPalletComponent
 import dev.t1r.themebuilder.feature.materialcolorspallet.MaterialColorsPalletComponent.Model
 import dev.t1r.themebuilder.feature.materialcolorspallet.store.MaterialColorsPalletStore.Intent
 import dev.t1r.themebuilder.feature.materialcolorspallet.store.MaterialColorsPalletStoreProvider
+import dev.t1r.themebuilder.repository.colors.material.MaterialColorsRepository
+import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
