+++ feature-material-colors-pallet/src/commonMain/kotlin/dev/t1r/themebuilder/feature/materialcolorspallet/store/MaterialColorsPalletStoreProvider.kt
@@ -5,10 +5,10 @@ import com.arkivanov.mvikotlin.core.store.Store
 import com.arkivanov.mvikotlin.core.store.StoreFactory
 import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineBootstrapper
 import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineExecutor
-import dev.t1r.themebuilder.data.colors.material.MaterialColorsRepository
-import dev.t1r.themebuilder.data.colors.theme.ThemeColorsRepository
 import dev.t1r.themebuilder.entity.colors.*
 import dev.t1r.themebuilder.feature.materialcolorspallet.store.MaterialColorsPalletStore.*
+import dev.t1r.themebuilder.repository.colors.material.MaterialColorsRepository
+import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
